+++ test/base64url_test.dart
@@ -10,78 +10,62 @@ import 'utils.dart';
 
 void main() {
   group('Test base64url', () {
+    test('encoding', () {
+      var inp = [0, 0, 0];
+      expect(toBase64(inp, url: true), "AAAA");
+    });
     test('encoding no padding', () {
       for (int i = 0; i < 100; ++i) {
         var b = randomBytes(i);
         var r = base64UrlEncode(b).replaceAll('=', '');
-        expect(toBase64Url(b, padding: false), r, reason: 'length $i');
+        expect(toBase64(b, url: true, padding: false), r, reason: 'length $i');
       }
     });
     test('decoding no padding', () {
       for (int i = 0; i < 100; ++i) {
         var b = randomBytes(i);
         var r = base64UrlEncode(b).replaceAll('=', '');
-        expect(fromBase64Url(r), equals(b), reason: 'length $i');
+        expect(fromBase64(r), equals(b), reason: 'length $i');
       }
     });
     test('encoding with padding', () {
       for (int i = 0; i < 100; ++i) {
         var b = randomBytes(i);
         var r = base64UrlEncode(b);
-        expect(toBase64Url(b, padding: true), r, reason: 'length $i');
+        expect(toBase64(b, url: true, padding: true), r, reason: 'length $i');
       }
     });
     test('decoding with padding', () {
       for (int i = 0; i < 100; ++i) {
         var b = randomBytes(i);
         var r = base64UrlEncode(b);
-        expect(fromBase64Url(r), equals(b), reason: 'length $i');
+        expect(fromBase64(r), equals(b), reason: 'length $i');
       }
     });
     test('encoding <-> decoding no padding', () {
       for (int i = 0; i < 100; ++i) {
         var b = randomBytes(i);
-        var r = toBase64Url(b, padding: false);
-        expect(fromBase64Url(r), equals(b), reason: 'length $i');
+        var r = toBase64(b, url: true, padding: false);
+        expect(fromBase64(r), equals(b), reason: 'length $i');
       }
     });
     test('encoding <-> decoding with padding', () {
       for (int i = 0; i < 100; ++i) {
         var b = randomBytes(i);
-        var r = toBase64Url(b, padding: true);
-        expect(fromBase64Url(r), equals(b), reason: 'length $i');
+        var r = toBase64(b, url: true, padding: true);
+        expect(fromBase64(r), equals(b), reason: 'length $i');
       }
     });
     group('decoding with invalid chars', () {
       test('"Hashlib!"', () {
-        expect(() => fromBase64Url("Hashlib!"), throwsFormatException);
+        expect(() => fromBase64("Hashlib!"), throwsFormatException);
       });
       test('"-10"', () {
-        expect(() => fromBase64Url(" 10"), throwsFormatException);
+        expect(() => fromBase64(" 10"), throwsFormatException);
       });
-      test('"s/mething"', () {
-        expect(() => fromBase64Url("s/mething"), throwsFormatException);
+      test('"s*mething"', () {
+        expect(() => fromBase64("s*mething"), throwsFormatException);
       });
     });
-    // group('decoding with invalid length', () {
-    //   test('"H"', () {
-    //     expect(() => fromBase64Url("H"), throwsFormatException);
-    //   });
-    //   test('"Ha"', () {
-    //     expect(() => fromBase64Url("Ha"), throwsFormatException);
-    //   });
-    //   test('"Has"', () {
-    //     expect(() => fromBase64Url("Has"), throwsFormatException);
-    //   });
-    //   test('"Hashl"', () {
-    //     expect(() => fromBase64Url("Hashl"), throwsFormatException);
-    //   });
-    //   test('"Hashli"', () {
-    //     expect(() => fromBase64Url("Hashli"), throwsFormatException);
-    //   });
-    //   test('"Hashlib"', () {
-    //     expect(() => fromBase64Url("Hashlib"), throwsFormatException);
-    //   });
-    // });
   });
 }