+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.0
+
+- Fixes padding issues with `base2` and `base16`
+- In `toBase32`, uses the parameter `lower` replacing `upper`.
+- Transfers `fromBase64Url` to `fromBase64` with extended alphabet.
+- Transfers `toBase64Url` to `toBase64` with optional `url` parameter.
+- Improves documentation
+
 # 1.0.0
 
 - First release